[parser] Adjust associativity to make PtrQuals nonterminal more useful
By making MULTI/REQUIRED/OPTIONAL/SINGLE right associative, we can get
rid of an unfortunate bit in our parser where we need to inline the
body of PtrQuals.

The downside here is that this breaks some stuff: currently
`{required multi := 1}` works, with `multi` being the name of
the property, but this change breaks that.